renamed slidePanel to slidePane
