+++ src/slidepane/createSlidePaneElement.ts
