+++ src/slidepane/example/index.html
