+++ src/slidepane/example/index.ts
