+++ src/slidepane/styles/slidePane.css
