+++ src/slidepane/styles/slidePane.css.d.ts
